requirements.txt
requests~=2.28.1
colorama~=0.4.5
psutil~=5.9.1
PyYAML~=6.0
boltons~=21.0.0
dateparser~=1.1.1
click~=8.1.3
importlib_metadata; python_version<"3.8"
backports.zoneinfo[tzdata];python_version<"3.9"
